16 bits keyboard

+++ Midiproject/src/avrMIDI.h
@@ -27,7 +27,7 @@ void Init_ports(void);
 void UART_out(uint8_t ch);
 void Init_UART0(uint16_t BAUDRATE);
 void MIDI_send(uint8_t command, uint8_t tone, uint8_t volume);
-uint8_t MIDI_Conversion(uint8_t pressed);
+uint8_t MIDI_Conversion(uint16_t pressed);
 
 void init_adc(void);
 void init_Timer0(void);
